refactor: 토큰 생성과 정보 추출 기능을 UseCase로 분리
