--- mohaeng/src/main/java/com/mohaeng/domain/authentication/service/JwtService.java
@@ -1,10 +1,0 @@
-package com.mohaeng.domain.authentication.service;
-
-import com.mohaeng.common.jwt.Claims;
-
-public interface JwtService {
-
-    String createAccessToken(final Claims claims);
-
-    Claims getClaims(final String token);
-}
