+++ mohaeng/src/main/java/com/mohaeng/infrastructure/authentication/jwt/usecase/CreateTokenUseCase.java
@@ -1,0 +1,13 @@
+package com.mohaeng.infrastructure.authentication.jwt.usecase;
+
+import com.mohaeng.common.jwt.Claims;
+
+public interface CreateTokenUseCase {
+
+    String command(final Command command);
+
+    record Command(
+            Claims claims
+    ) {
+    }
+}
