+++ mohaeng/src/test/java/com/mohaeng/infrastructure/authentication/jwt/service/ExtractClaimsTest.java
@@ -1,0 +1,50 @@
+package com.mohaeng.infrastructure.authentication.jwt.service;
+
+import com.mohaeng.common.jwt.Claims;
+import com.mohaeng.common.properties.JwtProperties;
+import com.mohaeng.infrastructure.authentication.jwt.exception.AccessTokenInvalidException;
+import com.mohaeng.infrastructure.authentication.jwt.usecase.ExtractClaimsUseCase;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+@DisplayName("ExtractClaims 은 ")
+class ExtractClaimsTest {
+
+    private static final String TOKEN =
+            "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJleHAiOjE2ODA2ODI4NTEsIm1lbWJlcklkIjoiMSJ9.s2E4VEA_w16a9Z9QxCSDtq8DNHD-VgRLRKluMA1frxZBEt6WERbrkAlNLYybF4-IH6s4Ogei52zSpEBq_LG9-g";
+    private final ExtractClaimsUseCase extractClaimsUseCase = new ExtractClaims(new MockJwtProperties());
+
+    @Test
+    @DisplayName("토큰을 받으면 해당 토큰의 클레임을 반환한다.")
+    void returnClaimsWhenGivenToken() {
+        // given
+        Claims returnClaims = extractClaimsUseCase.command(
+                new ExtractClaimsUseCase.Command(TOKEN)
+        );
+
+        // then
+        assertAll(
+                () -> assertThat(returnClaims.claims()).isNotEmpty()
+        );
+    }
+
+    @Test
+    @DisplayName("토큰이 올바르지 않다면 예외를 발생시킨다.")
+    void throwExceptionWhenInvalidJWT() {
+        // when, then
+        assertThatThrownBy(() -> extractClaimsUseCase.command(
+                        new ExtractClaimsUseCase.Command("invalid token")
+                )
+        ).isInstanceOf(AccessTokenInvalidException.class);
+    }
+
+    private static class MockJwtProperties extends JwtProperties {
+        public MockJwtProperties() {
+            super("7JWI64WV7ZWY7IS47JqU7KCA64qU7Iug64+Z7ZuI7J6F64uI64uk", 100L);
+        }
+    }
+}
