+++ mohaeng/src/test/java/com/mohaeng/infrastructure/authentication/jwt/service/MockJwtProperties.java
@@ -1,0 +1,9 @@
+package com.mohaeng.infrastructure.authentication.jwt.service;
+
+import com.mohaeng.common.properties.JwtProperties;
+
+public class MockJwtProperties extends JwtProperties {
+    public MockJwtProperties() {
+        super("7JWI64WV7ZWY7IS47JqU7KCA64qU7Iug64+Z7ZuI7J6F64uI64uk", 100L);
+    }
+}
